ENH: move ace clean to general
