+++ pysatNASA/instruments/ace_swepam_l2.py
@@ -64,8 +64,8 @@ _test_dates = {id: {tag: dt.datetime(2021, 1, 1) for tag in inst_ids[id]}
 # Use standard init routine
 init = functools.partial(mm_nasa.init, module=mm_ace, name=name)
 
-# Use default ace clean
-clean = mm_ace.clean
+# Use default clean
+clean = mm_nasa.clean
 
 # ----------------------------------------------------------------------------
 # Instrument functions
